Create the new 2bullet tower and rework health and damage system
the new tower has been created.
the health and damage system has been reworked to Floats instead of Ints

+++ Assets/Scripts/Enemies/HealthHandler.cs
@@ -6,16 +6,15 @@ public class HealthHandler : MonoBehaviour
 {
     [Header("Attributes")]
     [SerializeField] private int currencyWorth = 50;
-    public int hitPoints = 1;
+    public float hitPoints = 1f;
 
     private bool isDestroyed = false; 
 
-
-    public void TakeDamage(int dmg)
+    public void TakeDamage(float dmg)
     {
         hitPoints -= dmg;
 
-        if (hitPoints <= 0 && !isDestroyed)
+        if (hitPoints <= 0f && !isDestroyed)
         {
             EnemySpawner.onEnemyDestroy.Invoke();
             LevelManager.main.IncreaseCurrency(currencyWorth);
